requirements.txt
cycler==0.10.0
GPUtil==1.4.0
kiwisolver==1.0.1
matplotlib==3.0.2
numpy==1.15.4
opencv-contrib-python==3.4.5.20
opencv-python==3.4.4.19
Pillow==5.3.0
pyparsing==2.3.0
python-dateutil==2.7.5
PyYAML==3.13
scikit-learn==0.20.2
scipy==1.2.0
six==1.12.0
sklearn==0.0
torch==1.0.0
torchvision==0.2.1
selectivesearch
multiprocessing
apex   {	git clone https://github.com/NVIDIA/apex
		cd apex
		pip install -v --no-cache-dir --global-option="--cpp_ext" --global-option="--cuda_ext" .
       }
